chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.0.8",
    "langchain-google-genai>=3.1.0",
    "python-dotenv>=1.2.1",
]
